refactor: Nested file for actions
