+++ packages/backend/src/apps/frappe/actions/index.js
@@ -1,7 +1,7 @@
-import getDoc from './get-doc.js';
-import deleteDoc from './delete-doc.js';
-import createDoc from './create-doc.js';
-import getList from './get-list.js';
-import updateDoc from './update-doc.js';
+import getDoc from './get-doc/index.js';
+import deleteDoc from './delete-doc/index.js';
+import createDoc from './create-doc/index.js';
+import getList from './get-list/index.js';
+import updateDoc from './update-doc/index.js';
 
 export default [createDoc, getDoc, updateDoc, deleteDoc, getList];
